+++ crates/preset_env_base/Cargo.toml
@@ -10,9 +10,12 @@ version = "0.1.0"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+ahash = "0.7.4"
 anyhow = "1"
 browserslist-rs = "=0.8.0"
+dashmap = "4.0.2"
 from_variant = {version = "0.1.3", path = "../from_variant"}
+once_cell = "1.9.0"
 semver = {version = "1.0.4", features = ["serde"]}
 serde = {version = "1", features = ["derive"]}
 st-map = "0.1.2"
